Added first person player
First person player can walk, jump, and sense if they're on the ground. Need a method to toggle between overview camera and first person camera. Hanger is at 0,0 but it goes lower and the player won't fall below it. Either need to allow player to keep going down to need to move hangar up and make sure building works with that. Also need to change the layer of all other prefabs to "ground"